Everything works now, simplifying one thing
